ai_core/requirements.txt
Flask==2.2.3
Flask-RESTful==0.3.10
Flask-Cors==3.0.10
Flask-SocketIO==5.3.3
requests==2.28.2
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.0
joblib==1.2.0
scikit-learn==1.2.2
sentence-transformers==2.2.2
qdrant-client==1.3.0
langchain==0.0.191
transformers==4.28.1
yfinance==0.2.18
pycoingecko==3.1.0
pytest==7.3.1
pytest-mock==3.10.0
pytest-asyncio==0.21.0 # Added for async test support
psycopg2-binary==2.9.6
sqlalchemy==2.0.11
marshmallow==3.19.0
gunicorn==20.1.0
redis==4.5.5
websockets==11.0.2
